Landing subpages: align /demo, /privacy, /terms with the main landing (§4-b)
The main landing was redesigned (seal brand, floating-glass nav, oxblood,
Pretendard, card shadows) but the subpages lagged: no nav, seal-less footer,
and /demo used the full marketing hero scale.

- Extract StampMark (+ SEAL_STROKES) once to src/components/StampMark.tsx and
  import it from page.tsx and all three subpages — the SVG is no longer
  duplicated. Pure presentational component, works in server + client.
- /demo: add the floating-glass nav + seal footer; dial the intro heading down
  from the marketing .tagline scale to a section-scale .demo-title (no hero
  scale); give the acceptance/output lists (.outputs) quiet card rows so they
  match the .demo-card/.kv/.ev-*/.decision blocks (already on --shadow-card).
- /privacy + /terms: add the same nav + seal footer; restyle the .back link to
  the landing's quiet text-link treatment. Document-page density kept — no
  hero-scale type. Pretendard inherits from layout.tsx.

Skin only: no copy, route, or IA changes. Palette limited to existing
brand/gold/parchment tokens; no dark mode, emoji, gradient text, or new colors.
Gate green: pnpm turbo run typecheck lint build --filter @simsa/simsa-landing.
